% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_logical}
\alias{all_logical}
\title{Test that all data frame columns are logical}
\usage{
all_logical(.x)
}
\arguments{
\item{.x}{A data frame}
}
\value{
A logical indicating if all elements of .x are logical
}
\description{
Test that all data frame columns are logical
}
