% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2map.R
\name{core2map}
\alias{core2map}
\title{Create a data frame from flat TMA coordinates}
\usage{
core2map(x, row, col, val = NULL)
}
\arguments{
\item{x}{A data frame to use}

\item{row}{The column name for TMA row information}

\item{col}{The column name for TMA column information}

\item{val}{The column to display in each cell. If omitted (or NULL), generate a coordinate display \verb{(row,col)}.}
}
\value{
A data frame representing the row/col arrangement of the table
}
\description{
Create a data frame from flat TMA coordinates
}
