% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{equal_overlap}
\alias{equal_overlap}
\title{Compute overlap of equal values between two tables}
\usage{
equal_overlap(reference, target)
}
\arguments{
\item{reference}{A table representing the reference state to compare against}

\item{target}{Another table that may overlap in values with reference}
}
\value{
A percentage of equal values that overlap (number overlapping divided by total elements in reference)
}
\description{
Compute overlap of equal values between two tables
}
\examples{
\dontrun{
equal_overlap()
}
}
