% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{generate_alignment_mapping}
\alias{generate_alignment_mapping}
\title{Generating row/column mapping using an alignment function}
\usage{
generate_alignment_mapping(target, align_fun = alignment_functions[[1]])
}
\arguments{
\item{target}{The target(transformed) matrix to generate mappings}

\item{align_fun}{An alignment function from which mappings are used}
}
\value{
A table of source and target mappings
}
\description{
Create a mapping between a source data frame and the desired transform.
}
\details{
Assuming an alignment (rotation/reverse) is known, create a mapping between
coordinates in the src coordinate system (before rotation) into the reference system
defined by the transform.

The result is a data frame with the prefix of \code{src} for original coordinates (labels
and index) to the transformed (presumably corrected) coordinate system (labels and
index). For instance, the source row and columns indexes are \code{src_row_index} and
\code{src_col_index}. The corrected row/col index are \code{row_index} and \code{col_index}.
}
