% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{any_missing}
\alias{any_missing}
\title{Test if any elements of data frame are na}
\usage{
any_missing(.x)
}
\arguments{
\item{.x}{data frame}
}
\value{
A logical indicating if any elements are na
}
\description{
Test if any elements of data frame are na
}
\examples{
\dontrun{
any_missing(data.frame(A=c(1,2,3),B=c(2,NA,4)))
# returns [1] TRUE
}
}
